\name{SarsaTurns-package}
\alias{SarsaTurns-package}
\alias{SarsaTurns}
\docType{package}
\title{
\packageTitle{SarsaTurns}
}
\description{
\packageDescription{SarsaTurns}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{SarsaTurns}
\packageIndices{SarsaTurns}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{SarsaTurns}

Maintainer: \packageMaintainer{SarsaTurns}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
